examples/requirements.txt
betterproto[compiler]
pillow
requests
scipy
